refactor: switch `Makeswift`/REST API client to use runtime's `fetch`
