+++ .changeset/dry-suns-enjoy.md
@@ -1,0 +1,8 @@
+---
+'@makeswift/express-react': patch
+'@makeswift/react-router': patch
+'@makeswift/hono-react': patch
+'@makeswift/runtime': patch
+---
+
+refactor: switch `Makeswift` client to use runtime's `fetch` implementation
