+++ packages/express-react/src/client.ts
@@ -1,7 +1,1 @@
-import { type SiteVersion, MakeswiftClient } from '@makeswift/runtime/unstable-framework-support'
-
-export class Makeswift extends MakeswiftClient {
-  fetchOptions(_siteVersion: SiteVersion | null): Record<string, unknown> {
-    return {}
-  }
-}
+export { MakeswiftClient as Makeswift } from '@makeswift/runtime/unstable-framework-support'
