+++ packages/runtime/src/runtimes/react/testing/react-runtime.tsx
@@ -3,5 +3,11 @@ import { TestOrigins } from '../../../testing/fixtures'
 import { ReactRuntime } from '../react-runtime'
 
 export function createReactRuntime({ breakpoints }: { breakpoints?: BreakpointsInput } = {}) {
-  return new ReactRuntime({ fetch, breakpoints, ...TestOrigins })
+  return new ReactRuntime({
+    // Mock Service Worker patches global `fetch` for interception; make sure our test calls go
+    // through the patched version instead of an eagerly captured reference to the original
+    fetch: (...args) => globalThis.fetch(...args),
+    breakpoints,
+    ...TestOrigins,
+  })
 }
